refactor(trigger): extract telegram + verification to pure-Node @/lib
Foundation for the planned Trigger.dev removal — these two helpers had no
Trigger.dev coupling beyond their location. Moved to @/lib so they're
callable from any context (scheduler jobs, admin API routes, scripts).

- src/lib/telegram.ts (new, no behavior change)
- src/lib/scraper-verification.ts (new, swapped Gemini → DeepSeek per the
  consolidation goal in the local-rebuild plan)
- src/trigger/utils/telegram.ts is now a 1-line re-export shim
- src/trigger/verification.ts is now a 1-line re-export shim
- src/autoresearch/experiment-log.ts updated to import from @/lib/telegram

Both shims will be deleted in Phase 4 of the rebuild. No runtime behavior
change for callers.
